+++ app/src/main/java/com/rootstrap/android/ui/ui/theme/Color.kt
@@ -8,3 +8,5 @@ val NewLabelBlue = Color(0xFF2751B9)
 val BackgroundGray = Color(0xFFF4F7FA)
 val RichBlack = Color(0xFF1A1700)
 val BlueLink = Color(0xFF076CE0)
+val GreyTransparent = Color(0x44CCCCCC)
+val BlackTransparent = Color(0xAA000000)
